requirements.txt
click==6.6
facebook-sdk==2.0.0
Flask==0.11.1
Flask-SQLAlchemy==2.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
pkg-resources==0.0.0
psycopg2==2.6.2
requests==2.11.1
SQLAlchemy==1.1.3
Werkzeug==0.11.11
